[haproxy] sc send proper hostname
* Previously, the var `back_or_front` was not reset, so each loop could use the
  previous `back_or_front`.
* This `back_or_front` var is used to somehow determine the hostname for the
  service_check.
This problem is fixed.

Also:
* split mock and integration tests
